refactor operator utils
