+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/plan/nodes/calcite/Expand.scala
@@ -21,8 +21,8 @@ package org.apache.flink.table.plan.nodes.calcite
 import org.apache.calcite.plan.{RelOptCluster, RelOptCost, RelOptPlanner, RelTraitSet}
 import org.apache.calcite.rel.`type`.RelDataType
 import org.apache.calcite.rel.metadata.RelMetadataQuery
-import org.apache.calcite.rel.{RelNode, SingleRel}
-import org.apache.calcite.rex.{RexLiteral, RexNode}
+import org.apache.calcite.rel.{RelNode, RelWriter, SingleRel}
+import org.apache.calcite.rex.{RexInputRef, RexLiteral, RexNode}
 import org.apache.calcite.util.Litmus
 
 import java.util
@@ -79,6 +79,19 @@ abstract class Expand(
   }
 
   override def deriveRowType(): RelDataType = outputRowType
+
+  override def explainTerms(pw: RelWriter): RelWriter = {
+    val names = outputRowType.getFieldNames
+    val terms = projects.map {
+      project =>
+        project.zipWithIndex.map {
+          case (r: RexInputRef, i: Int) => s"${names.get(i)}=[${r.getName}]"
+          case (l: RexLiteral, i: Int) => s"${names.get(i)}=[${l.getValue3}]"
+          case (o, _) => s"$o"
+        }.mkString("{", ", ", "}")
+    }.mkString(", ")
+    super.explainTerms(pw).item("projects", terms)
+  }
 
   override def computeSelfCost(planner: RelOptPlanner, mq: RelMetadataQuery): RelOptCost = {
     val rowCnt = mq.getRowCount(this.getInput) * projects.size()
