+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/plan/nodes/logical/FlinkLogicalExpand.scala
@@ -22,14 +22,12 @@ import org.apache.flink.table.plan.nodes.FlinkConventions
 import org.apache.flink.table.plan.nodes.calcite.{Expand, LogicalExpand}
 
 import org.apache.calcite.plan.{Convention, RelOptCluster, RelOptRule, RelTraitSet}
+import org.apache.calcite.rel.RelNode
 import org.apache.calcite.rel.`type`.RelDataType
 import org.apache.calcite.rel.convert.ConverterRule
-import org.apache.calcite.rel.{RelNode, RelWriter}
-import org.apache.calcite.rex.{RexInputRef, RexLiteral, RexNode}
+import org.apache.calcite.rex.RexNode
 
 import java.util
-
-import scala.collection.JavaConversions._
 
 /**
   * Sub-class of [[Expand]] that is a relational expression
@@ -47,19 +45,6 @@ class FlinkLogicalExpand(
 
   override def copy(traitSet: RelTraitSet, inputs: util.List[RelNode]): RelNode = {
     new FlinkLogicalExpand(cluster, traitSet, inputs.get(0), outputRowType, projects, expandIdIndex)
-  }
-
-  override def explainTerms(pw: RelWriter): RelWriter = {
-    val names = outputRowType.getFieldNames
-    val terms = projects.map {
-      project =>
-        project.zipWithIndex.map {
-          case (r: RexInputRef, i: Int) => s"${names.get(i)}=[${r.getName}]"
-          case (l: RexLiteral, i: Int) => s"${names.get(i)}=[${l.getValue3}]"
-          case (o, _) => s"$o"
-        }.mkString("{", ", ", "}")
-    }.mkString(", ")
-    super.explainTerms(pw).item("projects", terms)
   }
 
 }
