+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/plan/nodes/logical/FlinkLogicalSort.scala
@@ -19,6 +19,7 @@
 package org.apache.flink.table.plan.nodes.logical
 
 import org.apache.flink.table.plan.nodes.FlinkConventions
+import org.apache.flink.table.plan.util.RelNodeUtil
 
 import org.apache.calcite.plan._
 import org.apache.calcite.rel.convert.ConverterRule
@@ -37,12 +38,12 @@ class FlinkLogicalSort(
     traits: RelTraitSet,
     child: RelNode,
     collation: RelCollation,
-    sortOffset: RexNode,
-    sortFetch: RexNode)
-  extends Sort(cluster, traits, child, collation, sortOffset, sortFetch)
+    offset: RexNode,
+    fetch: RexNode)
+  extends Sort(cluster, traits, child, collation, offset, fetch)
   with FlinkLogicalRel {
 
-  private lazy val limitStart: Long = if (offset != null) RexLiteral.intValue(offset) else 0L
+  private lazy val limitStart: Long = RelNodeUtil.getLimitStart(offset)
 
   override def copy(
       traitSet: RelTraitSet,
