+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/plan/nodes/physical/batch/BatchExecLimit.scala
@@ -19,12 +19,13 @@ package org.apache.flink.table.plan.nodes.physical.batch
 
 import org.apache.flink.table.plan.cost.FlinkCost._
 import org.apache.flink.table.plan.cost.FlinkCostFactory
+import org.apache.flink.table.plan.util.RelNodeUtil
 
 import org.apache.calcite.plan.{RelOptCluster, RelOptCost, RelOptPlanner, RelTraitSet}
 import org.apache.calcite.rel._
 import org.apache.calcite.rel.core.Sort
 import org.apache.calcite.rel.metadata.RelMetadataQuery
-import org.apache.calcite.rex.{RexLiteral, RexNode}
+import org.apache.calcite.rex.RexNode
 
 /**
   * Batch physical RelNode for [[Sort]].
@@ -35,24 +36,20 @@ class BatchExecLimit(
     cluster: RelOptCluster,
     traitSet: RelTraitSet,
     inputRel: RelNode,
-    limitOffset: RexNode,
-    limit: RexNode,
+    offset: RexNode,
+    fetch: RexNode,
     val isGlobal: Boolean)
   extends Sort(
     cluster,
     traitSet,
     inputRel,
     traitSet.getTrait(RelCollationTraitDef.INSTANCE),
-    limitOffset,
-    limit)
+    offset,
+    fetch)
   with BatchPhysicalRel {
 
-  private lazy val limitStart: Long = if (offset != null) RexLiteral.intValue(offset) else 0L
-  private lazy val limitEnd: Long = if (limit != null) {
-    RexLiteral.intValue(limit) + limitStart
-  } else {
-    Long.MaxValue
-  }
+  private lazy val limitStart: Long = RelNodeUtil.getLimitStart(offset)
+  private lazy val limitEnd: Long = RelNodeUtil.getLimitEnd(offset, fetch)
 
   override def copy(
       traitSet: RelTraitSet,
@@ -65,19 +62,9 @@ class BatchExecLimit(
 
   override def explainTerms(pw: RelWriter): RelWriter = {
     pw.input("input", getInput)
-      .item("offset", offsetToString)
-      .item("limit", limitToString)
+      .item("offset", limitStart)
+      .item("fetch", RelNodeUtil.fetchToString(fetch))
       .item("global", isGlobal)
-  }
-
-  private def offsetToString: String = s"$limitStart"
-
-  private def limitToString: String = {
-    if (limit != null) {
-      s"${RexLiteral.intValue(limit)}"
-    } else {
-      "unlimited"
-    }
   }
 
   override def computeSelfCost(planner: RelOptPlanner, mq: RelMetadataQuery): RelOptCost = {
